Initial version of fd-based sslproc.
This only includes a partial implementation of the helper binary and
does not yet include library bits.  This design uses two separate fd's
per connection: an "upper" appFd which the client application uses to
send requests and receive replies, and a "lower" rawFd on which the
helper sends requests for I/O and receives replies.