Checkpoint further work.
This collapses the session state down to a single file descriptor.  It
also adds MessageBuffer and MessageSocket helper classes to reduce
duplicated code in ControlSocket and SSLSession.  These classes should
also be useable in the library used by client applications.

The logic for dealing with kevents is also moved to KEvent.{h,cc}.

Note that this version does not use non-blocking I/O but instead uses
blocking I/O on UNIX datagram sockets that are assumed to reliably
deliver atomic messages.