Add limited error handling framework.
This mimics the OpenSSL error queue, but it only provides a single
error per thread rather than multiple errors.  Use this to provide
error reporting for PSSL_CTX_new.

Note that this depends on libcrypto's ERR_error_string() and adds
entries to libcrypto's table to report sslproc-specific errors as
duplicating all of that logic seems excessive.