Refactor error handling in messages.
Always include an error code (SSL_ERROR_*) explicitly in a result
message instead of depending on 'ret' and the command as context.  In
particular, sslproc needs to be able to return errors for malformed
messages for message types when 'ret' might not be defined as
returning an error.

This also permits adding message errors to the error queue in
waitForReply() centrally in the library.

While here, change 'ret' to a long to support SSL_CTX_ctrl() which
returns a long.