Support SSL message callbacks.
Save the local callback function pointer and argument directly in the
PSSL.  When callbacks are enabled or disabled, send a message to the
helper to enable disable callbacks.  When callbacks are enabled, the
helper installs a local callback which forwards the message callback
to the client via a SSLPROC_MSG_CB message.