Support servername callbacks (SSL_CTX_set_tlsext_servername_callback*).
This one is a bit funky as the callback is registered only on the
context but it is invoked on a session.  To faciliate this, new
context messages enable/disable the callback over the control socket
to the context.  However, when the local callback fires in the helper,
it needs to find the right SSL session and send a callback message
over the SSL session socket.  To handle this, the helper now keeps an
unordered_map<> to map local SSL * pointers to SSL sessions.