Add implementations of SSL_{get,set}_SSL_CTX.
SSL_set_SSL_CTX is a stub that just fails.  This could be supported if
sslproc's model changes to use a global sandbox instead of separate
sandboxes for each SSL_CTX.  In apache it is used to support switching
to a different cert for vhosts.  For now I think we can get by with
saying per-vhost certs don't work for the initial demo.