Add a simple PSSL_SESSION object.
This does not aim to implement the full API of SSL_SESSION, instead
enough state is cached locally to permit the implementation of
external session caching callbacks.  In particular, not all of the
fields are replicated, but the objects do include the ASN1
serialization of the helper's "real" SSL_SESSION object that can be
returned to the helper to construct a real SSL_SESSION object for the
session get callback.