Proxy callbacks for external session caching.
This is a bit complicated.  Nominally, an external verison of a cached
session (PSSL_SESSION) just needs to include the serialized format
(ASN1) of the internal session from the helper so that the helper's
session get callback can reconstruct a SSL_SESSION as well as the
unique ID used as a short name for sessions.

In addition we pass across a few other fields when creating an
external session to permit accessors used in Apache (time, compression
id).

The other wrinkle however is that we cannot construct an arbitrary
external session for the remove callback, but must instead pass an
existing, valid pointer to the session being removed.  To support
this, the client side SSL context includes a hash table indexed by the
unique ID, and when new external sessions are created by either the
new session or get session callbacks, they are inserted into the hash
table.  The remove callback message includes the unique ID which is
used to find the associated PSSL_SESSION in the hash table to be
passed to the application's remove callback.