Rework how helper callbacks find the active SSL session.
It is important that callbacks are invoked on a session that is
currently waiting for a reply as callbacks written to some other
session socket may not have a listening thread in the client.  Rather
than using an unordered_map of SSL to SSLSession to find the session
socket on which to post callback messages, add a global variable that
holds the "currently active" SSL session and write callback messages
to that session.

For now this assumes a single global session.  If we supported
multiple threads in the helper, this variable would simply have to be
per-thread instead.