Rework IPC channels to be thread-based.
This is more inline with how cocall works, and it also avoids some
potential confusion with finding the right socket on which to invoke
callbacks.

Instead of creating a new helper process and "control" socket for each
SSL_CTX and a "session" socket for each SSL, libsslproc now only
creates a single helper process used by all SSL_CTX's in a single
process.  A single "control" socket is still passed as fd 3 across
exec(), but the only messages the control socket now handles is
responding to NOP, and a new CREATE_COMMAND_SOCKET message which
accepts an fd.  This fd is used to create a new CommandSocket over
which all the other commands are sent and received.  libsslproc creates
a its own client-side CommandSocket for each thread the first time a
thread tries to send a message.

SSL and SSL_CTX objects are now assigned integer names ("targets").
The TargetStore class is a wrapper around unordered_map<> that maps
targets to a simple tuple containing a void * and the std::type_info
for the original object pointer.  The target values are allocated from
a unified namespace, but the desired object type is checked against
the stored std::type_info in the hash to reject mismatches.

Most Messages are now "targeted" messages that include a target in the
header that is the object on which to invoke the message.  NOP,
CREATE_COMMAND_SOCKET, and CREATE_CONTEXT are the only messages that
do not have a target.  Callbacks specify a target object as well, so
the client (libsslproc) also maintains a TargetStore populated by the
targets returned from the helper in response to CREATE_CONTEXT and
CREATE_SESSION.

Since SSL_CTX and SSL objects are no longer associated with an fd,
they are no longer implicitly freed by the fd getting EOF, instead
explicit FREE_CONTEXT and FREE_SESSION messages are sent by the
client when the client-side object is destroyed.

Another change bundled in this commit is to stop statically
instantiating MessageBuffer objects.  Instead, a MessageSocket
contains a std::stack of MessageBuffers and readMessage() allocates a
new message from the stack.  The MesssageBuffer is returned to the
caller by a MessageRef which is a unique_ptr<>-like wrapper class
which returns the MessageBuffer to the socket's stack once it is no
longer needed.  CommandSockets on both sides allocate a stack of 4
messages currently (a SWAG) to permit that many levels of recursion
(e.g. SSL_connect or SSL_accept can invoke callbacks that themselves
invoke other SSL_* methods).  If the stack is empty, readMessage()
fails and that failure eventaully bubbles back up.