Switch to SOCK_STREAM for the command socket.
Turn MessageSocket into an abstract class with child classes for
datagram and stream sockets.  Only expose a writeMessage() method that
supports control data in the datagram socket class.  Similarly, turn
readMessage() into a virtual method in MessageSocket.  In the stream
class this uses plain read() to first read the header to obtain the
length, and then does a second read() to read the payload.  The
datagram socket class continues to use recvmsg() with MSG_PEEK to read
the header before then reading the full data.

ControlSocket classes now inherit from the datagram socket class, and
CommandSocket inherits from the stream class.

LibMessageSocket made this a bit trickier on the library side.  It is
now a template mixin class that takes a datagram or socket class as
its templated base class.  To avoid moving all of LibMessageSocket
into the header, some of the methods that don't care about stream vs
datagram are now thin wrappers around static functions in a
LibMessageSocketHelpers namespace.