Use dedicated threads for each CommandSocket.
If the client application uses multiple threads it may make a request
from thread A which depends on work done by thread B, e.g. if thread A
reads from a SSL connection which calls back into thread A to read
from a BIO.  If thread B needs to use an SSL request to satisfy A, the
single-threaded helper would be stuck waiting for a reply from A and
never perform B's request.