Proxy SSL_CTX_set_verify() and its associated callback.
This one is a bit trickier as the callback isn't invoked on an SSL
directly, but is instead invoked on an X509_STORE_CTX object.  However,
the X509_STORE_CTX object contains a reference to the associated
SSL via the X509_STORE_CTX_get_ex_data() interface using the index
from SSL_get_ex_data_X509_STORE_CTX_idx().

To handle the callback, provide a wrapper for
SSL_get_ex_data_X509_STORE_CTX_idx() to allocate an index locally in
the client and pass enough state along in the callback to construct a
local X509_STORE_CTX object in the client to pass to the callback.
The callback returns the updated error value from the client-side
context so that the error can be updated in the helper context before
the helper callback returns.

Note that not all of the state of the X509_STORE_CTX is forwarded
across, but enough to satisfy the callback in Apache.  Notably, we
don't forward across the full certificate chain, only the current
certificate.