Add a stub for SSL_CTX_get_cert_store().
This method cannot be trivially wrapped as it returns a pointer to an
internal object that is not a full reference.  In addition, some
callers expect to be able to modify the objected pointed to resulting
in changes to the SSL_CTX to which the X509_STORE belongs.

For now, just always return nullptr.