Proxy SSL_CTX_{get,set}_client_CA_list().
The list of CA names is serialized into an array of individually
serialized structures that are converted back into an OpenSSL stack
(STACK_OF) on the other end.

Note that SSL_CTX_get_client_CA_list() returns a pointer to an
internal stack that the caller doesn't have an explicit reference to.
To avoid leaking memory, store the most recently fetched list in the
PSSL_CTX and free it either when the context is destroyed or when a
list is fetched.