Proxy SSL_set_verify().
Store a verify callback in PSSL and inherit it from PSSL_CTX in
PSSL_new().  When handling Message::VERIFY_CB, always use the callback
from the PSSL.  Note SSL_set_verify()'s semantics that a NULL callback
pointer doesn't clear the existing callback, but instead is used to
change the mode without altering the callback.  There is no way to
clear a verify callback on an SSL once it is set.