Rework the representation of PSSL_CIPHER objects.
Rather than being ephemeral objects, these need to be read-only
long-lived globals.  In libssl these are static global variables, and
Apache at least depends on the fact that the same pointer is always
returned for a given SSL_CIPHER.  In particular, it assumes that it
can search for the list of pointers returned by SSL_get_ciphers() to
see if it contains the pointer returned by SSL_get_current_cipher().

To support this, assign target IDs to SSL_CIPHERS in the helper
anytime a not-yet-seen cipher is returned by SSL_get_current_cipher()
or SSL_get_pending_cipher().  On the client side, a PSSL_CIPHER
objects are allocated the first time a cipher target id is seen.  The
PSSL_CIPHER is populated with cipher details via the new
Message::CIPHER_FETCH_INFO message when it is first created.  It then
remains read-only and is never freed.
PSSL_get_{current,pending}_cipher() return a pointer to the existing
PSSL_CIPHER if one exists for a given target.

This does remove the need for PSSL to hold static storage for
previously-returned PSSL_CIPHER objects.