Proxy SSL_get_peer_cert_chain().
This is another case where OpenSSL returns a pointer without a
reference.  It is made a bit worse in that we have to return the
cached pointer as mod_ssl expects to call it twice and get the same
pointer back (not just a semantically identical list), but the chain
can change if a session is reused.  Right now we'd return the old
chain as we don't invalidate the cached chain on reuse.