Proxy SSL_client_hello_get0_ext().
This API returns pointers to internal buffers, so PSSL keeps a list of
returned pointers to be free'd later.  This API really should only be
used in the client hello callback, so the list of pointers is freed
after the callback returns as well as when a session is deleted.