Handle SSL_get_session().
This one requires managing the lifecycle a bit of the object.
Previously, SSL_SESSION objects only had a valid target for the
lifetime of the "new" and "remove" session callbacks.  For
SSL_get_session, the Messsage::GET_SESSION always allocates a new
target for the session it returns.  On the library side, we assume
that the session doesn't change once it is non-NULL, so we create a
local PSSL_SESSION that we cache in the PSSL and free when the PSSL is
freed.  The target for this PSSL_SESSION is explicitly freed when the
PSSL is freed via a Message::SESSION_REMOVE_TARGET message.