First cut at supporting fork().
Install an pthread_atfork child handler (and export it from the
library for use in single-threaded applications) which tears down
the old command sockets and clears the TLS variable of the sole
thread in the new process.  This ensures new threads all allocate
new threads.

However, this model shares a single helper process among both the
parent and child process.  I think those semantics probably aren't
great as it allows a child process to change an SSL_CTX of the parent,
and different child processes can in theory access each other's
connections by forging target IDs.  I think instead that a more
faithful model would be for fork to allocate a new control socket
socketpair and send a FORK message to the helper.  The helper would
fork() and use the new socket as the control socket in the child.  The
application child would then close the old control socket fd and use
the new control socket fd instead.  This would mirror the semantics
today where the OpenSSL state is COW'd on fork and subsequent changes
by either parent or child would not affect the other.