Add Message::typeName().
This function returns the name of a Message::type value.

This cannot use a static array with designated initializers because
C++ doesn't permit things C added in the previous millenium.

Having to maintain two duplicate lists is a bit painful.  One could
refactor this with some macro that enumerates all the types passing
them as an argument to a caller-supplied macro if it is too painful in
the future.