Initial port to cocall/coaccept.
Mostly this entails adding MessageCoAccept and MessageCoCall classes
and using them as the base classes of ControlChannel and
CommandChannel in the helper and library.  Note that with sockets, the
Control channels were always datagram and the Command channels always
stream sockets.  However, with cocall/coaccept, the helper channels
are always coaccept and the library channels are always cocall.

In general, the coproc Message channel classes work by having message
writes save a copy of the message to write in a 'pendingWrite'
message.  The readMessage() methods call coaccept/cocall and pass the
pendingWrite along as the output buffer to be copied to the input
buffer of the other side.  For cocall, there should always be a
pendingWrite, but coaccept does not have a pendingWrite a helper
thread tries to read its first message.  When a pendingWrite is
available, the next message read is read back into that same buffer
avoiding the need to allocate a separate read buffer.

The FORK message and related machinery are not supported when built
with cocall/coaccept.